# tests/Api.Tests/Endpoints/OpenApi/OpenApiTests.OpenApi_VerifyAsExpected.verified.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Waste Obligations REST API",
    "description": "Manage obligation data in relation to EPR",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://localhost"
    }
  ],
  "paths": {
...
    "/compliance-declarations/unsubmitted": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search unsubmitted compliance declarations",
        "description": "Returns eligible organisations without a submitted or accepted compliance declaration",
        "operationId": "SearchUnsubmittedComplianceDeclarations",
        "parameters": [
          {
            "name": "obligationYear",
            "in": "query",
            "schema": {
              "maximum": 2050,
              "minimum": 2023,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "registrationType",
            "in": "query",
            "description": "Comma separated list of organisation registration type",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Case-insensitive partial match on organisation name or reference number",
            "schema": {
              "maxLength": 100,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma separated sort fields in priority order. Each field must use the format Field[asc] or Field[desc]. Fields: OrganisationName, OrganisationReferenceNumber, RecyclingObligations, PercentageMet",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-based), defaults to 1 if not specified",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number of items per page, defaults to 20 if not specified, max of 100",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnsubmittedOrganisationsPaged"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RegistrationType": {
        "enum": [
          "DirectProducer",
          "ComplianceScheme",
          null
        ]
      },
      "UnsubmittedOrganisation": {
        "required": [
          "name",
          "referenceNumber"
        ],
        "type": "object",
        "properties": {
          "organisationId": {
            "type": "string",
            "format": "uuid"
          },
          "obligationYear": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "registrationType": {
            "$ref": "#/components/schemas/RegistrationType"
          },
          "name": {
            "type": "string"
          },
          "referenceNumber": {
            "type": "string"
          },
          "recyclingObligationsMet": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "obligationCoveragePercentage": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "UnsubmittedOrganisationsPaged": {
        "type": "object",
        "properties": {
          "unsubmittedOrganisations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnsubmittedOrganisation"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}